disable malachite, reorg code and comments
