+++ zebra-crosslink/src/lib.rs
@@ -1,23 +1,33 @@
 //! Internal Zebra service for managing the Crosslink consensus protocol
 
-use multiaddr::Multiaddr;
-use rand::Rng;
-use rand::SeedableRng;
-use std::fmt;
-use std::str::FromStr;
-use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::broadcast;
-use tokio::sync::Mutex;
-use tokio::task::JoinHandle;
 use tokio::time::Instant;
-use tracing::{error, info, warn};
+use tracing::{info, warn};
+/*
+use rand::{CryptoRng, RngCore};
+use crate::core::Round as BFTRound;
+use std::str::FromStr;
+use multiaddr::Multiaddr;
+use rand::{SeedableRng, Rng};
+use malachitebft_app_channel::app::types::codec::Codec;
+use malachitebft_test::codec::proto::ProtobufCodec;
+use malachitebft_app_channel::AppMsg as BFTAppMsg;
+use async_trait::async_trait;
+use malachitebft_app_channel::app::events::*;
+use malachitebft_app_channel::app::types::core::*;
+use malachitebft_app_channel::app::types::*;
+use malachitebft_app_channel::app::*;
+use malachitebft_test::{
+    Address, Ed25519Provider, Genesis, Height as BFTHeight, PrivateKey, PublicKey, TestContext,
+    Validator, ValidatorSet,
+};
+*/
 
 pub mod service;
 
 use crate::service::{
-    ReadStateServiceProcedure, TFLServiceCalls, TFLServiceError, TFLServiceHandle,
-    TFLServiceRequest, TFLServiceResponse,
+    TFLServiceCalls, TFLServiceError, TFLServiceHandle, TFLServiceRequest, TFLServiceResponse,
 };
 
 // TODO: do we want to start differentiating BCHeight/PoWHeight, BFTHeight/PoSHeigh etc?
@@ -29,6 +39,7 @@ use zebra_state::{
 /// Placeholder activation height for Crosslink functionality
 pub const TFL_ACTIVATION_HEIGHT: BlockHeight = BlockHeight(2000);
 
+#[allow(dead_code)]
 #[derive(Debug)]
 pub(crate) struct TFLServiceInternal {
     val: u64,
@@ -131,6 +142,7 @@ async fn tfl_final_block_hash(call: &TFLServiceCalls) -> Option<BlockHash> {
     }
 }
 
+/*
 /// Make up a new value to propose
 /// A real application would have a more complex logic here,
 /// typically reaping transactions from a mempool and executing them against its state,
@@ -158,6 +170,7 @@ fn bft_make_value(
 
     malachitebft_test::Value { value, extensions }
 }
+*/
 
 const MAIN_LOOP_SLEEP_INTERVAL: Duration = Duration::from_millis(125);
 const MAIN_LOOP_INFO_DUMP_INTERVAL: Duration = Duration::from_millis(8000);
@@ -165,276 +178,277 @@ const MAIN_LOOP_INFO_DUMP_INTERVAL: Duration = Duration::from_millis(8000);
 async fn tfl_service_main_loop(internal_handle: TFLServiceHandle) -> Result<(), String> {
     let call = internal_handle.call.clone();
 
-    let mut rng = rand::rngs::StdRng::seed_from_u64(5);
-    let private_key = PrivateKey::generate(&mut rng);
-    let public_key = private_key.public_key();
-    let _address = Address::from_public_key(&public_key);
-    let _signing_provider = Ed25519Provider::new(private_key.clone());
-    let ctx = TestContext::new();
+    /*
+        let mut rng = rand::rngs::StdRng::seed_from_u64(5);
+        let private_key = PrivateKey::generate(&mut rng);
+        let public_key = private_key.public_key();
+        let _address = Address::from_public_key(&public_key);
+        let _signing_provider = Ed25519Provider::new(private_key.clone());
+        let ctx = TestContext::new();
 
-    let initial_validator_set = ValidatorSet::new(vec![Validator::new(public_key, 1)]);
-    let genesis = Genesis {
-        validator_set: initial_validator_set.clone(),
-    };
+        let initial_validator_set = ValidatorSet::new(vec![Validator::new(public_key, 1)]);
+        let genesis = Genesis {
+            validator_set: initial_validator_set.clone(),
+        };
 
-    let codec = ProtobufCodec;
+        let codec = ProtobufCodec;
 
-    let init_bft_height = BFTHeight::new(1);
-    let bft_node_handle = BFTNode {
-        private_key: private_key.clone(),
-    };
+        let init_bft_height = BFTHeight::new(1);
+        let bft_node_handle = BFTNode {
+            private_key: private_key.clone(),
+        };
 
-    // let mut bft_config = config::load_config(std::path::Path::new("C:\\Users\\azmre\\.malachite\\config\\config.toml"), None)
-    //     .expect("Failed to load configuration file");
-    let mut bft_config: malachitebft_test_cli::config::Config = Default::default(); // TODO: read from file?
-                                                                                    //
-    bft_config.consensus.p2p.listen_addr = Multiaddr::from_str("/ip4/127.0.0.1/tcp/12345").unwrap();
-    bft_config.consensus.p2p.discovery = config::DiscoveryConfig {
-        selector: config::Selector::from_str("random").unwrap(), // TODO: Kademlia?
-        bootstrap_protocol: config::BootstrapProtocol::from_str("full").unwrap(), // TODO: Kademlia?
-        num_outbound_peers: 1,                                   // TODO: 20
-        num_inbound_peers: 1,                                    // TODO: 20
-        ephemeral_connection_timeout: Duration::from_secs(5),    // TODO: ?
-        enabled: false,
-    };
+        // let mut bft_config = config::load_config(std::path::Path::new("C:\\Users\\azmre\\.malachite\\config\\config.toml"), None)
+        //     .expect("Failed to load configuration file");
+        let mut bft_config: malachitebft_test_cli::config::Config = Default::default(); // TODO: read from file?
+                                                                                        //
+        bft_config.consensus.p2p.listen_addr = Multiaddr::from_str("/ip4/127.0.0.1/tcp/12345").unwrap();
+        bft_config.consensus.p2p.discovery = config::DiscoveryConfig {
+            selector: config::Selector::from_str("random").unwrap(), // TODO: Kademlia?
+            bootstrap_protocol: config::BootstrapProtocol::from_str("full").unwrap(), // TODO: Kademlia?
+            num_outbound_peers: 1,                                   // TODO: 20
+            num_inbound_peers: 1,                                    // TODO: 20
+            ephemeral_connection_timeout: Duration::from_secs(5),    // TODO: ?
+            enabled: false,
+        };
 
-    bft_config.mempool.p2p.listen_addr = Multiaddr::from_str("/ip4/127.0.0.1/tcp/22345").unwrap();
-    bft_config.mempool.p2p.discovery = config::DiscoveryConfig {
-        selector: config::Selector::from_str("random").unwrap(), // TODO: Kademlia?
-        bootstrap_protocol: config::BootstrapProtocol::from_str("full").unwrap(), // TODO: Kademlia?
-        num_outbound_peers: 1,                                   // TODO: 20
-        num_inbound_peers: 1,                                    // TODO: 20
-        ephemeral_connection_timeout: Duration::from_secs(5),    // TODO: ?
-        enabled: true,
-    };
+        bft_config.mempool.p2p.listen_addr = Multiaddr::from_str("/ip4/127.0.0.1/tcp/22345").unwrap();
+        bft_config.mempool.p2p.discovery = config::DiscoveryConfig {
+            selector: config::Selector::from_str("random").unwrap(), // TODO: Kademlia?
+            bootstrap_protocol: config::BootstrapProtocol::from_str("full").unwrap(), // TODO: Kademlia?
+            num_outbound_peers: 1,                                   // TODO: 20
+            num_inbound_peers: 1,                                    // TODO: 20
+            ephemeral_connection_timeout: Duration::from_secs(5),    // TODO: ?
+            enabled: true,
+        };
 
-    info!(?bft_config);
+        info!(?bft_config);
 
-    let (mut channels, _engine_handle) = malachitebft_app_channel::start_engine(
-        ctx,
-        codec,
-        bft_node_handle,
-        bft_config,
-        Some(init_bft_height),
-        initial_validator_set,
-    )
-    .await
-    .unwrap();
+        let (mut channels, _engine_handle) = malachitebft_app_channel::start_engine(
+            ctx,
+            codec,
+            bft_node_handle,
+            bft_config,
+            Some(init_bft_height),
+            initial_validator_set,
+        )
+        .await
+        .unwrap();
+    */
 
-    // TODO mutable state here in order to correctly respond to messages.
-
-    let mut last_diagnostic_print = Instant::now();
     let mut run_instant = Instant::now();
-    let mut current_bft_height = init_bft_height;
-    let mut current_bft_round = Round::Nil;
-    let mut current_bft_proposer = None;
+    let mut last_diagnostic_print = Instant::now();
     let mut current_bc_tip = None;
     let mut current_bc_final = None;
-    let mut vote_extensions =
-        std::collections::HashMap::<BFTHeight, VoteExtensions<TestContext>>::new();
-    // let mut bft_values           = Vec::<Option<LocallyProposedValue>>::new();
-    loop {
-        // sleep if we are running ahead
 
-        let task_periodic_run = tokio::time::sleep_until(run_instant);
-        let task_new_malachite_event = channels.consensus.recv();
+    /*
+        // TODO mutable state here in order to correctly respond to messages.
+        let mut current_bft_height = init_bft_height;
+        let mut current_bft_round = Round::Nil;
+        let mut current_bft_proposer = None;
+        let mut vote_extensions =
+            std::collections::HashMap::<BFTHeight, VoteExtensions<TestContext>>::new();
+        // let mut bft_values           = Vec::<Option<LocallyProposedValue>>::new();
+    */
+    loop {
         tokio::select! {
-            _ = task_periodic_run => {
+            // sleep if we are running ahead
+            _ = tokio::time::sleep_until(run_instant) => {
                 run_instant += MAIN_LOOP_SLEEP_INTERVAL;
             }
-            ret = task_new_malachite_event => {
-                let msg = ret.expect("Channel to Malachite has been closed.");
-                match msg {
-                    // The first message to handle is the `ConsensusReady` message, signaling to the app
-                    // that Malachite is ready to start consensus
-                    BFTAppMsg::ConsensusReady { reply } => {
-                        info!("BFT Consensus is ready");
+        /* ret = channels.consensus.recv() => {
+            let msg = ret.expect("Channel to Malachite has been closed.");
+            match msg {
+                // The first message to handle is the `ConsensusReady` message, signaling to the app
+                // that Malachite is ready to start consensus
+                BFTAppMsg::ConsensusReady { reply } => {
+                    info!("BFT Consensus is ready");
 
-                        if reply.send(malachitebft_app_channel::ConsensusMsg::StartHeight(
-                                current_bft_height,
-                                genesis.validator_set.clone(),
-                        )).is_err() {
-                            error!("Failed to send ConsensusReady reply");
-                        }
-                    },
+                    if reply.send(malachitebft_app_channel::ConsensusMsg::StartHeight(
+                            current_bft_height,
+                            genesis.validator_set.clone(),
+                    )).is_err() {
+                        tracing::error!("Failed to send ConsensusReady reply");
+                    }
+                },
 
-                    // The next message to handle is the `StartRound` message, signaling to the app
-                    // that consensus has entered a new round (including the initial round 0)
-                    BFTAppMsg::StartedRound {
-                        height,
-                        round,
-                        proposer,
-                    } => {
-                        info!(%height, %round, %proposer, "Started round");
+                // The next message to handle is the `StartRound` message, signaling to the app
+                // that consensus has entered a new round (including the initial round 0)
+                BFTAppMsg::StartedRound {
+                    height,
+                    round,
+                    proposer,
+                } => {
+                    info!(%height, %round, %proposer, "Started round");
 
-                        current_bft_height   = height;
-                        current_bft_round    = round;
-                        current_bft_proposer = Some(proposer);
-                    },
+                    current_bft_height   = height;
+                    current_bft_round    = round;
+                    current_bft_proposer = Some(proposer);
+                },
 
-                    // At some point, we may end up being the proposer for that round, and the engine
-                    // will then ask us for a value to propose to the other validators.
-                    BFTAppMsg::GetValue {
-                        height,
-                        round,
-                        timeout: _,
-                        reply,
-                    } => {
-                        info!(%height, %round, "Consensus is requesting a value to propose");
+                // At some point, we may end up being the proposer for that round, and the engine
+                // will then ask us for a value to propose to the other validators.
+                BFTAppMsg::GetValue {
+                    height,
+                    round,
+                    timeout: _,
+                    reply,
+                } => {
+                    info!(%height, %round, "Consensus is requesting a value to propose");
 
-                        // TODO: read from existing
+                    // TODO: read from existing
 
-                        // // Check if we have a previously built value for that height and round
-                        // if let Some(proposal) = state.get_previously_built_value(height, round) {
-                        //     info!(value = %proposal.value.id(), "Re-using previously built value");
+                    // // Check if we have a previously built value for that height and round
+                    // if let Some(proposal) = state.get_previously_built_value(height, round) {
+                    //     info!(value = %proposal.value.id(), "Re-using previously built value");
 
-                        //     if reply.send(proposal).is_err() {
-                        //         error!("Failed to send GetValue reply");
-                        //     }
+                    //     if reply.send(proposal).is_err() {
+                    //         tracing::error!("Failed to send GetValue reply");
+                    //     }
 
-                        //     return Ok(());
-                        // }
+                    //     return Ok(());
+                    // }
 
-                        // // Otherwise, propose a new value
-                        // let proposal = state.propose_value(height, round);
-                        //
-
-                        tokio::time::sleep(Duration::from_millis(500)).await;
-
-                        let proposal = LocallyProposedValue::<TestContext> {
-                            value: bft_make_value(&mut rng, &mut vote_extensions, height, round),
-                            height,
-                            round,
-                        };
-
-                        // Send it to consensus
-                        if reply.send(proposal.clone()).is_err() {
-                            error!("Failed to send GetValue reply");
-                        }
-
-                        // // Decompose the proposal into proposal parts and stream them over the network
-                        // for stream_message in state.stream_proposal(proposal) {
-                        //     info!(%height, %round, "Streaming proposal part: {stream_message:?}");
-                        //     channels
-                        //         .network
-                        //         .send(malachitebft_app_channel::NetworkMsg::PublishProposalPart(stream_message))
-                        //         .await?;
-                        // }
-                    },
-
-                    BFTAppMsg::ProcessSyncedValue {
-                        height,
-                        round,
-                        proposer,
-                        value_bytes,
-                        reply,
-                    } => {
-                        info!(%height, %round, "Processing synced value");
-
-                        let value = codec.decode(value_bytes).unwrap();
-
-                        if reply.send(ProposedValue {
-                            height,
-                            round,
-                            valid_round: Round::Nil,
-                            proposer,
-                            value,
-                            validity: Validity::Valid,
-                            // extension: None, TODO? "does not have this field"
-                        }).is_err() {
-                            error!("Failed to send ProcessSyncedValue reply");
-                        }
-                    },
-
-                    // In some cases, e.g. to verify the signature of a vote received at a higher height
-                    // than the one we are at (e.g. because we are lagging behind a little bit),
-                    // the engine may ask us for the validator set at that height.
+                    // // Otherwise, propose a new value
+                    // let proposal = state.propose_value(height, round);
                     //
-                    // In our case, our validator set stays constant between heights so we can
-                    // send back the validator set found in our genesis state.
-                    BFTAppMsg::GetValidatorSet { height: _, reply } => {
-                        // TODO: parameterize by height
-                        if reply.send(genesis.validator_set.clone()).is_err() {
-                            error!("Failed to send GetValidatorSet reply");
-                        }
-                    },
 
-                    // After some time, consensus will finally reach a decision on the value
-                    // to commit for the current height, and will notify the application,
-                    // providing it with a commit certificate which contains the ID of the value
-                    // that was decided on as well as the set of commits for that value,
-                    // ie. the precommits together with their (aggregated) signatures.
-                    BFTAppMsg::Decided {
-                        certificate,
-                        extensions,
-                        reply,
-                    } => {
-                        info!(
-                            height = %certificate.height,
-                            round = %certificate.round,
-                            value = %certificate.value_id,
-                            "Consensus has decided on value"
-                        );
+                    tokio::time::sleep(Duration::from_millis(500)).await;
 
-                        // When that happens, we store the decided value in our store
-                        // TODO: state.commit(certificate, extensions).await?;
-                        current_bft_height = certificate.height.increment();
-                        current_bft_round  = Round::new(0);
+                    let proposal = LocallyProposedValue::<TestContext> {
+                        value: bft_make_value(&mut rng, &mut vote_extensions, height, round),
+                        height,
+                        round,
+                    };
 
-                        // And then we instruct consensus to start the next height
-                        if reply.send(malachitebft_app_channel::ConsensusMsg::StartHeight(
-                                current_bft_height,
-                                genesis.validator_set.clone(),
-                        )).is_err() {
-                            error!("Failed to send Decided reply");
-                        }
-                    },
+                    // Send it to consensus
+                    if reply.send(proposal.clone()).is_err() {
+                        tracing::error!("Failed to send GetValue reply");
+                    }
 
-                    BFTAppMsg::GetHistoryMinHeight { reply } => {
-                        // TODO: min height from DB
-                        let min_height = init_bft_height;
-                        info!("Reply min_height: {}", min_height);
-                        if reply.send(min_height).is_err() {
-                            error!("Failed to send GetHistoryMinHeight reply");
-                        }
-                    },
+                    // // Decompose the proposal into proposal parts and stream them over the network
+                    // for stream_message in state.stream_proposal(proposal) {
+                    //     info!(%height, %round, "Streaming proposal part: {stream_message:?}");
+                    //     channels
+                    //         .network
+                    //         .send(malachitebft_app_channel::NetworkMsg::PublishProposalPart(stream_message))
+                    //         .await?;
+                    // }
+                },
 
-                    BFTAppMsg::GetDecidedValue { height, reply } => {
-                        let decided_value = None; // state.get_decided_value(&height).cloned();
+                BFTAppMsg::ProcessSyncedValue {
+                    height,
+                    round,
+                    proposer,
+                    value_bytes,
+                    reply,
+                } => {
+                    info!(%height, %round, "Processing synced value");
 
-                        if reply.send(decided_value).is_err() {
-                            error!("Failed to send GetDecidedValue reply");
-                        }
-                    },
+                    let value = codec.decode(value_bytes).unwrap();
 
-                    BFTAppMsg::ExtendVote {
-                        height: _,
-                        round: _,
-                        value_id: _,
-                        reply,
-                    } => {
-                        // TODO
-                        if reply.send(None).is_err() {
-                            error!("Failed to send ExtendVote reply");
-                        }
-                    },
+                    if reply.send(ProposedValue {
+                        height,
+                        round,
+                        valid_round: Round::Nil,
+                        proposer,
+                        value,
+                        validity: Validity::Valid,
+                        // extension: None, TODO? "does not have this field"
+                    }).is_err() {
+                        tracing::error!("Failed to send ProcessSyncedValue reply");
+                    }
+                },
 
-                    BFTAppMsg::VerifyVoteExtension {
-                        height: _,
-                        round: _,
-                        value_id: _,
-                        extension: _,
-                        reply,
-                    } => {
-                        // TODO
-                        if reply.send(Ok(())).is_err() {
-                            error!("Failed to send VerifyVoteExtension reply");
-                        }
-                    },
+                // In some cases, e.g. to verify the signature of a vote received at a higher height
+                // than the one we are at (e.g. because we are lagging behind a little bit),
+                // the engine may ask us for the validator set at that height.
+                //
+                // In our case, our validator set stays constant between heights so we can
+                // send back the validator set found in our genesis state.
+                BFTAppMsg::GetValidatorSet { height: _, reply } => {
+                    // TODO: parameterize by height
+                    if reply.send(genesis.validator_set.clone()).is_err() {
+                        tracing::error!("Failed to send GetValidatorSet reply");
+                    }
+                },
 
-                    _ => error!(?msg, "Unhandled message from Malachite"),
-                }
+                // After some time, consensus will finally reach a decision on the value
+                // to commit for the current height, and will notify the application,
+                // providing it with a commit certificate which contains the ID of the value
+                // that was decided on as well as the set of commits for that value,
+                // ie. the precommits together with their (aggregated) signatures.
+                BFTAppMsg::Decided {
+                    certificate,
+                    extensions,
+                    reply,
+                } => {
+                    info!(
+                        height = %certificate.height,
+                        round = %certificate.round,
+                        value = %certificate.value_id,
+                        "Consensus has decided on value"
+                    );
+
+                    // When that happens, we store the decided value in our store
+                    // TODO: state.commit(certificate, extensions).await?;
+                    current_bft_height = certificate.height.increment();
+                    current_bft_round  = Round::new(0);
+
+                    // And then we instruct consensus to start the next height
+                    if reply.send(malachitebft_app_channel::ConsensusMsg::StartHeight(
+                            current_bft_height,
+                            genesis.validator_set.clone(),
+                    )).is_err() {
+                        tracing::error!("Failed to send Decided reply");
+                    }
+                },
+
+                BFTAppMsg::GetHistoryMinHeight { reply } => {
+                    // TODO: min height from DB
+                    let min_height = init_bft_height;
+                    info!("Reply min_height: {}", min_height);
+                    if reply.send(min_height).is_err() {
+                        tracing::error!("Failed to send GetHistoryMinHeight reply");
+                    }
+                },
+
+                BFTAppMsg::GetDecidedValue { height, reply } => {
+                    let decided_value = None; // state.get_decided_value(&height).cloned();
+
+                    if reply.send(decided_value).is_err() {
+                        tracing::error!("Failed to send GetDecidedValue reply");
+                    }
+                },
+
+                BFTAppMsg::ExtendVote {
+                    height: _,
+                    round: _,
+                    value_id: _,
+                    reply,
+                } => {
+                    // TODO
+                    if reply.send(None).is_err() {
+                        tracing::error!("Failed to send ExtendVote reply");
+                    }
+                },
+
+                BFTAppMsg::VerifyVoteExtension {
+                    height: _,
+                    round: _,
+                    value_id: _,
+                    extension: _,
+                    reply,
+                } => {
+                    // TODO
+                    if reply.send(Ok(())).is_err() {
+                        tracing::error!("Failed to send VerifyVoteExtension reply");
+                    }
+                },
+
+                _ => tracing::error!(?msg, "Unhandled message from Malachite"),
             }
+        }*/
         }
 
         let new_bc_tip = if let Ok(ReadStateResponse::Tip(val)) =
@@ -454,20 +468,19 @@ async fn tfl_service_main_loop(internal_handle: TFLServiceHandle) -> Result<(), 
             current_bc_final
         };
 
-        #[allow(unused_mut)]
         let mut internal = internal_handle.internal.lock().await;
         // from this point onwards we must race to completion in order to avoid stalling incoming requests
 
         if new_bc_final != current_bc_final {
             info!("final changed to {:?}", new_bc_final);
             if let Some(hash) = new_bc_final {
-                internal.final_change_tx.send(hash);
+                let _ = internal.final_change_tx.send(hash);
             }
             // TODO: walk difference in height & send all
         }
 
         if !internal.is_tfl_activated {
-            if let Some((height, hash)) = new_bc_tip {
+            if let Some((height, _hash)) = new_bc_tip {
                 if height < TFL_ACTIVATION_HEIGHT {
                     continue;
                 } else {
@@ -500,21 +513,7 @@ async fn tfl_service_main_loop(internal_handle: TFLServiceHandle) -> Result<(), 
     }
 }
 
-use crate::core::Round as BFTRound;
-use async_trait::async_trait;
-use malachitebft_app_channel::app::events::*;
-use malachitebft_app_channel::app::types::codec::Codec;
-use malachitebft_app_channel::app::types::core::*;
-use malachitebft_app_channel::app::types::*;
-use malachitebft_app_channel::app::*;
-use malachitebft_app_channel::AppMsg as BFTAppMsg;
-use malachitebft_test::codec::proto::ProtobufCodec;
-use malachitebft_test::{
-    Address, Ed25519Provider, Genesis, Height as BFTHeight, PrivateKey, PublicKey, TestContext,
-    Validator, ValidatorSet,
-};
-use rand::{CryptoRng, RngCore};
-
+/*
 struct BFTNode {
     private_key: PrivateKey,
 }
@@ -602,6 +601,7 @@ impl Node for BFTNode {
         Ok(())
     }
 }
+*/
 
 async fn tfl_service_incoming_request(
     internal_handle: TFLServiceHandle,
@@ -615,12 +615,6 @@ async fn tfl_service_incoming_request(
 
     #[allow(unreachable_patterns)]
     match request {
-        TFLServiceRequest::IncrementVal => {
-            let ret = internal.val;
-            internal.val += 1;
-            Ok(TFLServiceResponse::ValIncremented(ret))
-        }
-
         TFLServiceRequest::FinalBlockHash => Ok(TFLServiceResponse::FinalBlockHash(
             tfl_final_block_hash(&internal_handle.call).await,
         )),
@@ -740,34 +734,3 @@ async fn tfl_service_incoming_request(
         _ => Err(TFLServiceError::NotImplemented),
     }
 }
-
-pub fn spawn_new_tfl_service(
-    read_state_service_call: ReadStateServiceProcedure,
-) -> (TFLServiceHandle, JoinHandle<Result<(), String>>) {
-    let handle1 = TFLServiceHandle {
-        internal: Arc::new(Mutex::new(TFLServiceInternal {
-            val: 0,
-            latest_final_height: BlockHeight(0),
-            latest_final_hash: BlockHash([0_u8; 32]),
-            is_tfl_activated: false,
-            final_change_tx: broadcast::channel(16).0,
-        })),
-        call: TFLServiceCalls {
-            read_state: read_state_service_call,
-        },
-    };
-    let handle2 = handle1.clone();
-
-    (
-        handle1,
-        tokio::spawn(async move { tfl_service_main_loop(handle2).await }),
-    )
-}
-
-impl fmt::Display for TFLServiceError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "TFLServiceError: {:?}", self)
-    }
-}
-
-impl std::error::Error for TFLServiceError {}
